refactor: エンジン停止後quitを呼ぶ前にhandleを呼ぶようにした

+++ src/background.ts
@@ -923,31 +923,18 @@ app.on("before-quit", async (event) => {
   const numLivingEngineProcess = Object.entries(killingProcessPromises).length;
 
   // すべてのエンジンプロセスが停止している
-  if (numLivingEngineProcess === 0) {
-    if (appState.willRestart || vvppManager.hasMarkedEngineDirs()) {
+  if (numLivingEngineProcess === 0 && !vvppManager.hasMarkedEngineDirs()) {
+    if (appState.willRestart) {
+      // 再起動フラグが立っている場合はフラグを戻して再起動する
       log.info(
-        "All ENGINE processes are killed, running post engine kill process"
+        "Post engine kill process done. Now restarting app because of willRestart flag"
       );
-      // awaitする前にevent.preventDefault()を呼び出さないとアプリがそのまま終了してしまう
       event.preventDefault();
 
-      // エンジン終了後の処理を実行
-      await vvppManager.handleMarkedEngineDirs();
+      appState.willRestart = false;
+      appState.willQuit = false;
 
-      if (appState.willRestart) {
-        // 再起動フラグが立っている場合はフラグを戻して再起動する
-        log.info(
-          "Post engine kill process done. Now restarting app because of willRestart flag"
-        );
-
-        appState.willRestart = false;
-        appState.willQuit = false;
-
-        start();
-      } else {
-        // アプリケーションの終了をを再試行する
-        app.quit();
-      }
+      start();
     } else {
       log.info("Post engine kill process done. Now quit app");
     }
@@ -983,10 +970,14 @@ app.on("before-quit", async (event) => {
   // すべてのエンジンプロセスキル処理が完了するまで待機
   await Promise.all(waitingKilledPromises);
 
-  // アプリケーションの終了を再試行する
+  // エンジン終了後の処理を実行
   log.info(
-    "All ENGINE process kill operations done. Attempting to quit app again"
+    "All ENGINE process kill operations done. running post engine kill process"
   );
+  await vvppManager.handleMarkedEngineDirs();
+
+  // アプリケーションの終了を再試行する
+  log.info("Attempting to quit app again");
   app.quit();
   return;
 });
